+++ spree/core/spec/models/spree/gift_card_batch_spec.rb
@@ -20,7 +20,9 @@ RSpec.describe Spree::GiftCardBatch, type: :model do
         expect(record).to receive(:publish_event).with('gift_card_batch.updated')
         allow(record).to receive(:publish_event).with(anything)
 
-        record.touch
+        Timecop.travel(1.minute.from_now) do
+          record.update_attribute(:updated_at, Time.current)
+        end
       end
     end