+++ spree/core/spec/models/spree/return_authorization_spec.rb
@@ -31,7 +31,9 @@ describe Spree::ReturnAuthorization, type: :model do
         expect(record).to receive(:publish_event).with('return_authorization.updated')
         allow(record).to receive(:publish_event).with(anything)
 
-        record.touch
+        Timecop.travel(1.minute.from_now) do
+          record.update_attribute(:updated_at, Time.current)
+        end
       end
     end